pyproject.toml
[project]
name = "edgetx-themes-tools"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
	"pyyaml",
	"rich",
	"pillow",
]

[tool.ruff]
line-length = 100
